Cast long long to int64_t initially rather than on return so that we don't internally deal with any long long types.
[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 3454]

+++ src/libbiomeval/be_io_properties.cpp
@@ -217,11 +217,11 @@ BiometricEvaluation::IO::Properties::getPropertyAsInteger(
 	}
 
 	/* Convert the string value to integer */
-	long long conVal = strtoll(value.c_str(), NULL, base);
+	int64_t conVal = (int64_t)strtoll(value.c_str(), NULL, base);
 	if (errno == ERANGE)
 		throw Error::ConversionError("Value out of range");
 
-	return ((int64_t)conVal);
+	return (conVal);
 }
 
 double
